Migrate individual plotting tests out of core
